+++ SDIS2/src/peer/RMIStub.java
@@ -36,7 +36,7 @@ public interface RMIStub extends Remote {
      * @param reservedSpace reserved space
      * @throws RemoteException  exception
      */
-    void reclaimProtocol(int reservedSpace) throws RemoteException;
+    void reclaimProtocol(long reservedSpace) throws RemoteException;
 
     /**
      * Interface method to call the State Protocol class
